+++ tiledb/sm/filesystem/s3.cc
@@ -714,7 +714,12 @@ Status S3::object_size(const URI& uri, uint64_t* nbytes) const {
 }
 
 Status S3::read(
-    const URI& uri, off_t offset, void* buffer, uint64_t length) const {
+    const URI& uri,
+    const off_t offset,
+    void* const buffer,
+    const uint64_t length,
+    const uint64_t read_ahead_length,
+    uint64_t* const length_returned) const {
   RETURN_NOT_OK(init_client());
 
   if (!uri.is_s3()) {
@@ -726,14 +731,17 @@ Status S3::read(
   Aws::S3::Model::GetObjectRequest get_object_request;
   get_object_request.WithBucket(aws_uri.GetAuthority())
       .WithKey(aws_uri.GetPath());
-  get_object_request.SetRange(("bytes=" + std::to_string(offset) + "-" +
-                               std::to_string(offset + length - 1))
-                                  .c_str());
-  get_object_request.SetResponseStreamFactory([buffer, length]() {
-    auto streamBuf = new boost::interprocess::bufferbuf((char*)buffer, length);
-    return Aws::New<Aws::IOStream>(
-        constants::s3_allocation_tag.c_str(), streamBuf);
-  });
+  get_object_request.SetRange(
+      ("bytes=" + std::to_string(offset) + "-" +
+       std::to_string(offset + length + read_ahead_length - 1))
+          .c_str());
+  get_object_request.SetResponseStreamFactory(
+      [buffer, length, read_ahead_length]() {
+        auto streamBuf = new boost::interprocess::bufferbuf(
+            (char*)buffer, length + read_ahead_length);
+        return Aws::New<Aws::IOStream>(
+            constants::s3_allocation_tag.c_str(), streamBuf);
+      });
 
   auto get_object_outcome = client_->GetObject(get_object_request);
   if (!get_object_outcome.IsSuccess()) {
@@ -741,7 +749,10 @@ Status S3::read(
         std::string("Failed to read S3 object ") + uri.c_str() +
         outcome_error_message(get_object_outcome)));
   }
-  if ((uint64_t)get_object_outcome.GetResult().GetContentLength() != length) {
+
+  *length_returned =
+      static_cast<uint64_t>(get_object_outcome.GetResult().GetContentLength());
+  if (*length_returned < length) {
     return LOG_STATUS(Status::S3Error(
         std::string("Read operation returned different size of bytes.")));
   }